Modified Models to include status in DeploymentTarget and gave RunningDeployment related_name and related_query_name Made migrations to reflect those changes Add SessionAuthentication to Token authentication_classes Handle createToken api request coming from deployment_manager and logged in as admin Add csrf token to requests made to createToken for both RoomDemoComponent and DeploymentManager Handle updated coming from RDM through sockets Establish WebSocket communcation from Deployment Manager and Verboze Handle Updates coming from Verboze through websockets Update js-api-utils

+++ deployment_manager/consumers.py
@@ -1,21 +1,21 @@
 from api.models import Token as VerbozeToken
-from deployment_manager.models import RemoteDeploymentMachine, Repository, DeploymentTarget, Firmware
-from deployment_manager.serializers import RepositorySerializer
-from channels import Channel
+from deployment_manager.models import RemoteDeploymentMachine, Repository, DeploymentTarget, Firmware, Deployment, RunningDeployment
+from deployment_manager.serializers import RepositorySerializer, RunningDeploymentSerializer
+from channels import Channel, Group
 from django.db.models import Q
 from django.utils import timezone
 from django.core.exceptions import ValidationError
 from django.contrib.contenttypes.models import ContentType
+from django.contrib.auth import get_user_model
 
 import json
 
 def get_valid_token(token):
     # delete old tokens
     VerbozeToken.objects.filter(~Q(expiry=None), expiry__lt=timezone.now()).delete()
+
     try:
-        # only allow tokens created for remote deployment machines
-        rdm_contenttype = ContentType.objects.get(model='remotedeploymentmachine')
-        token_object = VerbozeToken.objects.get(id=token, content_type=rdm_contenttype)
+        token_object = VerbozeToken.objects.get(id=token)
     except (VerbozeToken.DoesNotExist, ValidationError):
         return None
     return token_object
@@ -23,40 +23,62 @@ def get_valid_token(token):
 def send_repo_data(message):
     repos = Repository.objects.all()
     serializer = RepositorySerializer(repos, many=True)
-    message.reply_channel.send({'text': json.dumps({'repos': serializer.data})})
+    message.reply_channel.send({"text": json.dumps({"repos": serializer.data})})
+
+def send_updated_running_deployments():
+    running_deployments = RunningDeployment.objects.all().order_by("-id")
+    serializer = RunningDeploymentSerializer(running_deployments, many=True)
+    Group("updates_from_rdms").send({"text": json.dumps({"running_deployments": serializer.data})})
 
 def update_rdm_deployment_targets(rdm, message_content):
     deployment_targets = message_content["deployment_targets"]
+    if type(deployment_targets) == list: # validate right format
 
-    # delete old deployment targets
-    rdm.targets.all().delete()
+        # delete old deployment targets
+        rdm.targets.all().delete()
 
-    for target in deployment_targets:
-        DeploymentTarget.objects.create(
-            remote_deployment_machine=rdm,
-            identifier=target['identifier']
-        )
+        for target in deployment_targets:
+            new_target_identifier = target.get("identifier")
+            new_target_status = target.get("status")
+            if new_target_identifier and new_target_status: # validate identifier/status provided
+                DeploymentTarget.objects.create(
+                    remote_deployment_machine=rdm,
+                    identifier=new_target_identifier,
+                    status=new_target_status
+                )
 
 def update_rdm_firmwares(rdm, message_content):
     firmwares = message_content["firmwares"]
+    if type(firmwares) == list: # validate right format
 
-    # delete old firmwares
-    rdm.firmwares.all().delete()
+        # delete old firmwares
+        rdm.firmwares.all().delete()
 
-    for firmware in firmwares:
-        Firmware.objects.create(
-            remote_deployment_machine=rdm,
-            name=firmware['name']
-        )
+        for firmware in firmwares:
+            new_firmware_name = firmware.get("name")
+            if new_firmware_name: # validate name provided
+                Firmware.objects.create(
+                    remote_deployment_machine=rdm,
+                    name=new_firmware_name
+                )
 
 def update_running_deployment_target_stdout(rdm, message_content):
-    pass
+    deployment_id = message_content["deployment_update"].get("deployment")
+    updated_stdout = message_content["deployment_update"].get("message")
+    if deployment_id and updated_stdout: # validate right format
+        try:
+            deployment = Deployment.objects.get(id=deployment_id)
+            running_deployment = deployment.running_deployments.first()
+            running_deployment.stdout += updated_stdout
+            running_deployment.save()
+        except Exception:
+            pass
 
 def ws_connect(message, token):
     token_object = get_valid_token(token)
     # making sure RDM object exists for token
     if token_object and isinstance(token_object.content_object, RemoteDeploymentMachine):
-        message.reply_channel.send({'accept': True})
+        message.reply_channel.send({"accept": True})
 
         # send repo information for client to clone/pull latest
         send_repo_data(message)
@@ -65,8 +87,19 @@ def ws_connect(message, token):
         rdm = token_object.content_object
         rdm.channel_name = message.reply_channel
         rdm.save()
+    elif token_object and isinstance(token_object.content_object, get_user_model()):
+        # only accept connections from superusers
+        if token_object.content_object.is_superuser:
+            message.reply_channel.send({"accept": True})
+
+            # create Channel Group for updates to be sent
+            Group("updates_from_rdms").add(message.reply_channel)
+            send_updated_running_deployments()
+
+        else:
+            message.reply_channel.send({"accept": False})
     else:
-        message.reply_channel.send({'accept': False})
+        message.reply_channel.send({"accept": False})
 
 def ws_receive(message, token):
     token_object = get_valid_token(token)
@@ -79,9 +112,9 @@ def ws_receive(message, token):
             update_rdm_deployment_targets(rdm, message_content)
         elif message_content.get("firmwares"):
             update_rdm_firmwares(rdm, message_content)
-        elif message_content.get("running_deployment_target"):
-            print(message_content)
+        elif message_content.get("deployment_update"):
             update_running_deployment_target_stdout(rdm, message_content)
+            send_updated_running_deployments()
         else: # ignore message if not any of the above
             pass
     else:
@@ -92,4 +125,8 @@ def ws_disconnect(message, token):
     if token_object and isinstance(token_object.content_object, RemoteDeploymentMachine):
         token_object.content_object.delete()
         token_object.delete()
-
+    elif token_object and isinstance(token_object.content_object, get_user_model()):
+        # remove reply_channel from group
+        Group("updates_from_rdms").discard(message.reply_channel)
+        # delete token
+        token_object.delete()
